+++ api/router.proto
@@ -17,9 +17,9 @@ syntax = "proto3";
 import "github.com/gogo/protobuf/gogoproto/gogo.proto";
 import "google/protobuf/empty.proto";
 
-package router;
+package ttn.lorawan.v2;
 
-option go_package = "go.thethings.network/lorawan-stack-legacy/router";
+option go_package = "go.thethings.network/lorawan-stack-legacy/pkg/ttnpb";
 
 message UplinkMessage {
   bytes payload = 1;
