--- discovery/discovery.proto
@@ -1,39 +1,0 @@
-// Copyright © 2018 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-
-import "github.com/gogo/protobuf/gogoproto/gogo.proto";
-
-package discovery;
-
-option go_package = "go.thethings.network/lorawan-stack-legacy/discovery";
-
-message GetRequest {
-  string id = 1 [(gogoproto.customname) = "ID"];
-  string service_name = 2;
-}
-
-message Announcement {
-  string id = 1 [(gogoproto.customname) = "ID"];
-  string service_name = 2;
-  string service_version = 3;
-  string net_address = 11;
-  string public_key  = 12;
-  string certificate = 13;
-}
-
-service Discovery {
-  rpc Get(GetRequest) returns (Announcement);
-}
