+++ pkg/ttnpb/announcement.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package discovery
+package ttnpb
 
 import (
 	"crypto/tls"
